servisleri denemek için resources klasörü altında static altında front end yazıldı. Çeşitli düzenlemeler yapıldı. Product için unit test denemesi yapıldı.

+++ ProductService/src/main/java/com/javaexample/ProductService/Service/Concretes/ProductServiceImpl.java
@@ -4,6 +4,8 @@ import com.javaexample.ProductService.Entity.Product;
 import com.javaexample.ProductService.Kafka.Producer.ElasticSearch;
 import com.javaexample.ProductService.Repository.ProductRepository;
 import com.javaexample.ProductService.Service.Abstracts.ProductService;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -23,13 +25,18 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public void deleteProduct(int id) {
+        Optional<Product> product = productRepository.findById(id);
+        if (product.isEmpty()) {
+            throw new RuntimeException("Product not found");
+        }
         productRepository.deleteById(id);
     }
 
     @Override
     public Product createProduct(Product product) {
-        elasticSearchProducer.sendMessage(product);
-        return productRepository.save(product);
+        Product savedProduct = productRepository.save(product); // Önce kaydedin
+        elasticSearchProducer.sendMessage(savedProduct); // Sonra mesajı gönderin
+        return savedProduct;
     }
 
     @Override
@@ -40,13 +47,32 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public List<Product> getAllProduct() {
-        return productRepository.findAll();
+    public Page<Product> getAllProduct(Pageable pageable) {
+        return productRepository.findAll(pageable);
     }
+
+
 
     @Override
     public Optional<Product> getProductById(int id) {
         return productRepository.findById(id);
+    }
+
+    @Override
+    public Integer getProductStockCount(int productId) {
+        return productRepository.stockCount(productId);
+    }
+
+    @Override
+    public void decrementStock(int productId) {
+        Product product = productRepository.findById(productId).orElseThrow(() -> new RuntimeException("Product not found"));
+
+        if (product.getStock() <= 0) {
+            throw new RuntimeException("Stock is already zero");
+        }
+
+        product.setStock(product.getStock() - 1);
+        productRepository.save(product);
     }
